Upgraded the plugin code to Strapi v5
