+++ .gitignore
@@ -1,3 +1,5 @@
 /node_modules
 /dist
 /build
+.DS_Store
+.env
