+++ admin/custom.d.ts
@@ -1,0 +1,2 @@
+declare module '@strapi/design-system/*';
+declare module '@strapi/design-system';
