+++ admin/src/components/Initializer.tsx
@@ -1,11 +1,6 @@
-/**
- *
- * Initializer
- *
- */
-
 import { useEffect, useRef } from 'react';
-import pluginId from '../../pluginId';
+
+import { PLUGIN_ID } from '../pluginId';
 
 type InitializerProps = {
   setPlugin: (id: string) => void;
@@ -15,10 +10,10 @@ const Initializer = ({ setPlugin }: InitializerProps) => {
   const ref = useRef(setPlugin);
 
   useEffect(() => {
-    ref.current(pluginId);
+    ref.current(PLUGIN_ID);
   }, []);
 
   return null;
 };
 
-export default Initializer;
+export { Initializer };
