+++ admin/src/components/LocationInput.tsx
@@ -5,137 +5,119 @@
  */
 
 import {
-  Box,
-  Button,
-  Grid,
-  GridItem,
-  ModalBody,
-  ModalFooter,
-  ModalHeader,
-  ModalLayout,
-  Typography,
-} from "@strapi/design-system";
-import L from "leaflet";
-import "leaflet/dist/leaflet.css";
-import React, { useEffect, useMemo, useRef, useState } from "react";
-//@ts-ignore
-import markerIcon from "leaflet/dist/images/marker-icon.png";
-//@ts-ignore
-import iconRetina from "leaflet/dist/images/marker-icon-2x.png";
-//@ts-ignore
-import _ from "lodash";
-import { MapContainer, Marker, TileLayer, useMap } from "react-leaflet";
-import LocationInputForm from "./LocationInputForm";
-import LocationTextInput from "./LocationTextInput";
-
-//@ts-ignore
-const icon = L.icon({
-  iconUrl: markerIcon,
-  iconRetinaUrl: iconRetina,
-  iconSize: [25, 41],
-  iconAnchor: [12.5, 41],
-});
-
-const parseValue = (value: string): [number | null, number | null] => {
-  try {
-    const object = JSON.parse(value);
-
-    if (!object?.lat || !object?.lng) {
+    Box,
+    Button,
+    Grid,
+    Modal,
+    Typography,
+  } from "@strapi/design-system";
+  import L from "leaflet";
+  import "leaflet/dist/leaflet.css";
+  import "../leaflet.css";
+  import { useEffect, useMemo, useRef, useState } from "react";
+  //@ts-ignore
+  import markerIcon from "leaflet/dist/images/marker-icon.png";
+  //@ts-ignore
+  import iconRetina from "leaflet/dist/images/marker-icon-2x.png";
+  //@ts-ignore
+  import _ from "lodash";
+  import { MapContainer, Marker, TileLayer, useMap } from "react-leaflet";
+  import LocationInputForm from "./LocationInputForm";
+  import LocationTextInput from "./LocationTextInput";
+  
+  //@ts-ignore
+  const icon = L.icon({
+    iconUrl: markerIcon,
+    iconRetinaUrl: iconRetina,
+    iconSize: [25, 41],
+    iconAnchor: [12.5, 41],
+  });
+  
+  const parseValue = (value: { lat: number | null, lng: number | null }): [number | null, number | null] => {
+    try {
+      if (!value?.lat || !value?.lng) {
+        return [null, null];
+      }
+      return [
+        _.pick(value, ["lat", "lng"]).lat,
+        _.pick(value, ["lat", "lng"]).lng,
+      ];
+    } catch (error) {
       return [null, null];
     }
-    return [
-      _.pick(object, ["lat", "lng"]).lat,
-      _.pick(object, ["lat", "lng"]).lng,
-    ];
-  } catch (error) {
-    return [null, null];
-  }
-};
-
-//@ts-ignore
-const LocationInput = ({ value, onChange, name, attribute }) => {
-  const [defLat, defLng] = [49.195678016117164, 16.608182539182483];
-  const [[lat, lng], setLocation] = useState(parseValue(value));
-  const [isModalVisible, setIsModalVisible] = useState(false);
-
-  function FlyMapTo() {
-    const map = useMap();
-
-    useEffect(() => {
-      map.setView([lat ? lat : defLat, lng ? lng : defLng], 15);
-    }, [lat, lng]);
-
-    return null;
-  }
-
-  const markerRef = useRef(null);
-
-  const eventHandlers = useMemo(
-    () => ({
-      dragend() {
-        const marker = markerRef.current;
-        if (marker != null) {
-          //@ts-ignore
-          const { lat: newLat, lng: newLng } = marker.getLatLng();
-          handleSetLocation([newLat, newLng]);
-        }
-      },
-    }),
-    []
-  );
-
-  const handleSetLocation = (newValue: [number | null, number | null]) => {
-    setLocation(newValue);
-    onChange({
-      target: {
-        name,
-        value: JSON.stringify({ lat: newValue[0], lng: newValue[1] }),
-        type: attribute.type,
-      },
-    });
   };
-
-  return (
-    <Box>
-      <Typography fontWeight="bold" variant="pi">
-        {name}
-      </Typography>
-      <Grid gap={5}>
-        <LocationInputForm
-          lat={lat}
-          lng={lng}
-          handleSetLocation={handleSetLocation}
-        />
-        <GridItem col={12}>
-          <Button onClick={() => setIsModalVisible((prev) => !prev)}>
-            Open map
-          </Button>
-          {isModalVisible && (
-            <ModalLayout
-              onClose={() => setIsModalVisible((prev) => !prev)}
-              labelledBy="title"
-            >
-              <ModalHeader>
-                <Typography
-                  fontWeight="bold"
-                  textColor="neutral800"
-                  as="h2"
-                  id="title"
-                >
-                  Title
-                </Typography>
-              </ModalHeader>
-              <ModalBody>
-                <Grid gap={5} className="pb-2">
-                  <LocationInputForm
-                    lat={lat}
-                    lng={lng}
-                    handleSetLocation={handleSetLocation}
-                  />
-                </Grid>
-                <LocationTextInput handleSetLocation={handleSetLocation} />
-                <Box paddingTop={6}>
-                  <MapContainer
+  
+  //@ts-ignore
+  const LocationInput = ({ value, onChange, name, attribute }) => {
+    const [defLat, defLng] = [49.195678016117164, 16.608182539182483];
+    const [[lat, lng], setLocation] = useState(parseValue(value));
+  
+    function FlyMapTo() {
+      const map = useMap();
+  
+      useEffect(() => {
+        map.setView([lat ? lat : defLat, lng ? lng : defLng], 15);
+      }, [lat, lng]);
+  
+      return null;
+    }
+  
+    const markerRef = useRef(null);
+  
+    const eventHandlers = useMemo(
+      () => ({
+        dragend() {
+          const marker = markerRef.current;
+          if (marker != null) {
+            //@ts-ignore
+            const { lat: newLat, lng: newLng } = marker.getLatLng();
+            handleSetLocation([newLat, newLng]);
+          }
+        },
+      }),
+      []
+    );
+  
+    const handleSetLocation = (newValue: [number | null, number | null]) => {
+      setLocation(newValue);
+      onChange({
+        target: {
+          name,
+          value: JSON.stringify({ lat: newValue[0], lng: newValue[1] }),
+          type: attribute.type,
+        },
+      });
+    };
+  
+    return (
+      <Box>
+        <Typography fontWeight="bold" variant="pi">
+          {name}
+        </Typography>
+        <Grid.Root gap={5}>
+          <LocationInputForm
+            lat={lat}
+            lng={lng}
+            handleSetLocation={handleSetLocation}
+            displayingInModal={false}
+          />
+          <Grid.Item col={12}>
+            <Modal.Root>
+              <Modal.Trigger>
+                <Button>Open Map View</Button>
+              </Modal.Trigger>
+              <Modal.Content>
+                <Modal.Header><Modal.Title>Map View</Modal.Title></Modal.Header>
+                <Modal.Body>
+                    <LocationInputForm
+                        lat={lat}
+                        lng={lng}
+                        handleSetLocation={handleSetLocation}
+                        displayingInModal={true}
+                      />
+                    <LocationTextInput handleSetLocation={handleSetLocation} />
+                    <Box paddingTop={6}>
+                    <MapContainer
                     center={[lat ? lat : defLat, lng ? lng : defLng]}
                     zoom={12}
                     scrollWheelZoom={false}
@@ -154,23 +136,21 @@ const LocationInput = ({ value, onChange, name, attribute }) => {
                     ></Marker>
                     <FlyMapTo />
                   </MapContainer>
-                </Box>
-              </ModalBody>
-              <ModalFooter
-                endActions={
-                  <>
-                    <Button onClick={() => setIsModalVisible((prev) => !prev)}>
-                      Ok
-                    </Button>
-                  </>
-                }
-              />
-            </ModalLayout>
-          )}
-        </GridItem>
-      </Grid>
-    </Box>
-  );
-};
-
-export default LocationInput;
+                    </Box>
+                </Modal.Body>
+                
+                <Modal.Footer>
+                <Modal.Close>
+                  <Button>OK</Button>
+                </Modal.Close>
+                </Modal.Footer>
+              </Modal.Content>
+            </Modal.Root>
+          </Grid.Item>
+        </Grid.Root>
+      </Box>
+    );
+  };
+  
+  export default LocationInput;
+
