+++ admin/src/components/PluginIcon.tsx
@@ -1,0 +1,5 @@
+import { PinMap } from '@strapi/icons';
+
+const PluginIcon = () => <PinMap />;
+
+export { PluginIcon };
