+++ admin/src/index.ts
@@ -1,0 +1,52 @@
+import { getTranslation } from './utils/getTranslation';
+import { PLUGIN_ID } from './pluginId';
+import { Initializer } from './components/Initializer';
+import { PluginIcon } from './components/PluginIcon';
+
+export default {
+  register(app: any) {
+    
+
+    app.registerPlugin({
+      id: PLUGIN_ID,
+      initializer: Initializer,
+      isReady: false,
+      name: PLUGIN_ID,
+    });
+
+    app.customFields.register({
+      name: "location",
+      pluginId: PLUGIN_ID, 
+      type: "json", 
+      intlLabel: {
+        id: `${PLUGIN_ID}.location.label`,
+        defaultMessage: "Location",
+      },
+      intlDescription: {
+        id: `${PLUGIN_ID}.location.description`,
+        defaultMessage: "Select any location",
+      },
+      icon: PluginIcon,
+      components: {
+        Input: async () =>
+          import(
+            /* webpackChunkName: "input-component" */ "./components/LocationInput"
+          ),
+      },
+    });
+  },
+
+  async registerTrads({ locales }: { locales: string[] }) {
+    return Promise.all(
+      locales.map(async (locale) => {
+        try {
+          const { default: data } = await import(`./translations/${locale}.json`);
+
+          return { data, locale };
+        } catch {
+          return { data: {}, locale };
+        }
+      })
+    );
+  },
+};
