+++ admin/src/leaflet.css
@@ -1,0 +1,4 @@
+.leaflet-container {
+    height: 800px;
+}
+
