--- admin/src/utils/getTrad.ts
@@ -1,5 +1,0 @@
-import pluginId from '../pluginId';
-
-const getTrad = (id: string) => `${pluginId}.${id}`;
-
-export default getTrad;
