--- custom.d.ts
@@ -1,6 +1,0 @@
-declare module "@strapi/design-system/*";
-declare module "@strapi/design-system";
-declare module "@strapi/icons";
-declare module "@strapi/icons/*";
-declare module "@strapi/helper-plugin";
-declare module "*.png";
