+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@notum-cz/strapi-plugin-location",
-  "version": "1.3.3",
+  "version": "2.0.1",
   "description": "This plugin allows users to create location inputs and store latitude and longitude values as geometry types in a PostGIS database. It also provides functionality to filter items based on their location.",
   "keywords": [
     "strapi",
@@ -21,38 +21,69 @@
   "maintainers": [
     "Notum Technologies s.r.o. <sales@notum.cz> (https://notum.cz/en/strapi)",
     "Dominik Juriga <dominik.juriga@notum.cz> (https://notum.cz/en/strapi)"
+  ],  
+  "type": "commonjs",
+  "exports": {
+    "./package.json": "./package.json",
+    "./strapi-admin": {
+      "types": "./dist/admin/src/index.d.ts",
+      "source": "./admin/src/index.ts",
+      "import": "./dist/admin/index.mjs",
+      "require": "./dist/admin/index.js",
+      "default": "./dist/admin/index.js"
+    },
+    "./strapi-server": {
+      "types": "./dist/server/src/index.d.ts",
+      "source": "./server/src/index.ts",
+      "import": "./dist/server/index.mjs",
+      "require": "./dist/server/index.js",
+      "default": "./dist/server/index.js"
+    }
+  },
+  "files": [
+    "dist"
   ],
   "scripts": {
-    "develop": "NODE_ENV=development tsc -p tsconfig.server.json -w"
+    "build": "strapi-plugin build",
+    "watch": "strapi-plugin watch",
+    "watch:link": "strapi-plugin watch:link",
+    "verify": "strapi-plugin verify",
+    "test:ts:front": "run -T tsc -p admin/tsconfig.json",
+    "test:ts:back": "run -T tsc -p server/tsconfig.json"
   },
   "dependencies": {
-    "@notum-cz/strapi-plugin-location": "^1.3.2",
-    "@strapi/design-system": "^1.19.0",
-    "@strapi/helper-plugin": "^4.25.11",
-    "@strapi/icons": "^1.19.0",
+    "@strapi/design-system": "^2.0.0-rc.30",
+    "@strapi/icons": "^2.0.0-rc.30",
     "leaflet": "^1.9.4",
-    "prop-types": "^15.7.2",
-    "react-leaflet": "^4.2.1"
+    "react-intl": "^7.1.14",
+    "react-leaflet": "4.2.1"
   },
   "devDependencies": {
-    "@strapi/strapi": "^4.25.1",
-    "@strapi/typescript-utils": "^4.24.5",
-    "@types/leaflet": "^1.9.12",
-    "@types/react": "^18.3.6",
-    "@types/react-dom": "^18.3.0",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/styled-components": "^5.1.26",
+    "@strapi/sdk-plugin": "^5.3.2",
+    "@strapi/strapi": "^5.31.0",
+    "@strapi/typescript-utils": "^5.31.0",
+    "@types/leaflet": "^1.9.21",
+    "@types/minimatch": "^5.1.2",
+    "@types/react": "^19.2.6",
+    "@types/react-dom": "^19.2.3",
+    "prettier": "^3.6.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^5.3.4",
-    "styled-components": "^5.3.11",
-    "typescript": "5.6.2"
+    "react-leaflet": "4.2.1",
+    "react-router-dom": "^6.30.2",
+    "styled-components": "^6.1.19",
+    "typescript": "^5.9.3"
   },
   "peerDependencies": {
-    "@strapi/strapi": "4.x"
+    "@strapi/sdk-plugin": "^5.3.2",
+    "@strapi/strapi": "^5.31.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.30.2",
+    "styled-components": "^6.1.19"
   },
   "engines": {
-    "node": ">=14.19.1 <=20.x.x",
+    "node": ">=18.0.0 <=22.x.x",
     "npm": ">=6.0.0"
   },
   "strapi": {
