--- server/controllers/index.ts
@@ -1,5 +1,0 @@
-import myController from './my-controller';
-
-export default {
-  myController,
-};
