--- server/controllers/my-controller.ts
@@ -1,10 +1,0 @@
-import { Strapi } from '@strapi/strapi';
-
-export default ({ strapi }: { strapi: Strapi }) => ({
-  index(ctx) {
-    ctx.body = strapi
-      .plugin('location-plugin')
-      .service('myService')
-      .getWelcomeMessage();
-  },
-});
