--- server/destroy.ts
@@ -1,3 +1,0 @@
-import { Strapi } from "@strapi/strapi";
-
-export default async ({ strapi }: { strapi: Strapi }) => {};
