+++ server/src/destroy.ts
@@ -1,0 +1,7 @@
+import type { Core } from '@strapi/strapi';
+
+const destroy = ({ strapi }: { strapi: Core.Strapi }) => {
+  // destroy phase
+};
+
+export default destroy;
