+++ server/src/index.ts
@@ -1,12 +1,19 @@
-import register from './register';
+/**
+ * Application methods
+ */
 import bootstrap from './bootstrap';
 import destroy from './destroy';
+import register from './register';
+
+/**
+ * Plugin server methods
+ */
 import config from './config';
 import contentTypes from './content-types';
 import controllers from './controllers';
-import routes from './routes';
 import middlewares from './middlewares';
 import policies from './policies';
+import routes from './routes';
 import services from './services';
 
 export default {
