+++ server/src/routes/index.ts
@@ -1,0 +1,3 @@
+const routes = {};
+
+export default routes;
