--- strapi-server.js
@@ -1,2 +1,0 @@
-"use strict";
-module.exports = require("./dist/server");
