--- tsconfig.json
@@ -1,37 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "./",
-    "outDir": "build",
-    "target": "esnext",
-    "module": "commonjs",
-    "jsx": "react",
-    "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "removeComments": true,
-    "sourceMap": true,
-    "declaration": true,
-    "declarationMap": true,
-    "downlevelIteration": true,
-    "incremental": true,
-    "allowJs": true,
-    "skipLibCheck": true,
-    "resolveJsonModule": true,
-    "noImplicitAny": false,
-  },
-
-  "include": [    "./server/**/*",
-  "./admin/**/*",
-  "./types/**/*",
-  "./strapi-server.*",
-  "./strapi-admin.*",
-   "custom.d.ts",
-  "./admin/**/*",
-  "./setup-package.ts",],
-
-  "exclude": [
-    "**/node_modules/**/*",
-    "**/build/**/*",
-    "**/dist/**/*"
-  ]
-}
